// src/payload-types.ts
export interface Media {
  id: string;
  user: string | User;
  alt: string;
  caption?: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    card?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    tablet?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "users".
 */
export interface User {
  id: string;
  name?: string | null;
  balance: number;
  role: 'admin' | 'enterprise' | 'professional' | 'member' | 'user';
  domain?: string | null;
  members?: (string | User)[] | null;
  associatedWith?: string | null;
  rappAccess?:
    | {
        rappId: string | PrivateRapp;
        getAccess?: ('read' | 'delete' | 'create' | 'update')[] | null;
        id?: string | null;
      }[]
    | null;
  publicRapps?: (string | PublicRapp)[] | null;
  privateRapps?: (string | PrivateRapp)[] | null;
  updatedAt: string;
  createdAt: string;
  enableAPIKey?: boolean | null;
  apiKey?: string | null;
  apiKeyIndex?: string | null;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  password?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "privateRapps".
 */
export interface PrivateRapp {
  id: string;
  type: 'text' | 'image';
  model: string | Model;
  key: 'test' | 'prod';
  systemprompt?: string | null;
  prompt: string;
  negativeprompt?: string | null;
  status: 'pending' | 'approved' | 'denied';
  slug?: string | null;
  imageinput: boolean;
  name: string;
  description: string;
  priceapplicable: boolean;
  price?: number | null;
  totalCost: number;
  images?:
    | {
        image: string | Media;
        id?: string | null;
      }[]
    | null;
  creator: string | User;
  systemVariables?:
    | {
        name: string;
        identifier: string;
        displayName: string;
        description: string;
        placeholder?: string | null;
        type: 'string' | 'number' | 'boolean' | 'select';
        allowMultiple: boolean;
        options?: string[] | null;
        id?: string | null;
      }[]
    | null;
  promptVariables?:
    | {
        name: string;
        identifier: string;
        displayName: string;
        description: string;
        placeholder?: string | null;
        type: 'string' | 'number' | 'boolean' | 'select';
        allowMultiple: boolean;
        options?: string[] | null;
        id?: string | null;
      }[]
    | null;
  negativeVariables?:
    | {
        name: string;
        identifier: string;
        displayName: string;
        description: string;
        placeholder?: string | null;
        type: 'string' | 'number' | 'boolean' | 'select';
        allowMultiple: boolean;
        options?: string[] | null;
        id?: string | null;
      }[]
    | null;
  access?:
    | {
        userId: string | User;
        getAccess?: ('read' | 'delete' | 'update')[] | null;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "models".
 */
export interface Model {
  id: string;
  name: string;
  description: string;
  about: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  type: 'text' | 'image';
  imageinput: boolean;
  provider: 'groq' | 'openai' | 'replicate';
  settings: {
    name: string;
    type: 'integer' | 'string' | 'boolean' | 'select';
    description: string;
    options?: string[] | null;
    allowMultiple: boolean;
    id?: string | null;
  }[];
  systemprompt: boolean;
  negativeprompt: boolean;
  prodkeys?: {
    groq?: {
      modelname: string;
      apikey: string;
    };
    replicate?: {
      modelname: string;
      apitoken: string;
    };
    openai?: {
      modelname: string;
      apikey: string;
    };
  };
  testkeys?: {
    groq?: {
      modelname: string;
      apikey: string;
    };
    replicate?: {
      modelname: string;
      apitoken: string;
    };
    openai?: {
      modelname: string;
      apikey: string;
    };
  };
  cost: number;
  commissionapplicable: boolean;
  commission?: number | null;
  examples: {
    example: string;
    id?: string | null;
  }[];
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "publicRapps".
 */
export interface PublicRapp {
  id: string;
  type: 'text' | 'image';
  model: string | Model;
  computationcost?: number | null;
  commission?: number | null;
  systemprompt?: string | null;
  prompt: string;
  negativeprompt?: string | null;
  status: 'pending' | 'approved' | 'denied';
  needsApproval: boolean;
  featured: boolean;
  imageinput: boolean;
  likes?:
    | {
        user: string | User;
        id?: string | null;
      }[]
    | null;
  name: string;
  description: string;
  price: number;
  totalCost?: number | null;
  images?:
    | {
        image: string | Media;
        id?: string | null;
      }[]
    | null;
  creator: string | User;
  updatedAt: string;
  createdAt: string;
}